Added option to display owners and all orgs
Report now can output the owners of an org with the --owner option. If
you're an owner of that org, you see all owners, otherwise only public
owners.

If you add the --email option, owners and their emails are displayed.

If you use the --all-my-orgs, the default is all orgs for which you have
owner permissions.