Repo2docker (#63)
Support repo2docker to "freeze" versions.

This project has been a mess for a while in terms of python versions and GitHub3.py versions.

Repo2Docker looks to restore (my) sanity, by allowing me to have different docker images for the currently active set of working items.

This initial setup is bare bones, but creates a viable docker image where both *.ipynb and scripts can be executed. That did require providing another way to pass in a GitHub PAT, which is now done.