Switch to ipywidgets data entry for email
This change minimizes the chance of committing the notebook with PII in
it.

This moves the entered text to an "output" section of the cell, which
must be removed before commit (enforced by pre-commit filters).

It also makes the workbook easier to use, as you no longer have to worry
about messing up the python string formatting.