ACL listing output (#77)
Crawl all repos looking for potential ACL in config files.

This approach has a lot of false positives, and the results need to be interpreted by someone with domain knowledge.

The "final output" is in CSV format -- by copy/paste into gSheets, a more reviewable view can be constructed.

There are heuristic ignore lists for orgs, repos, and paths-in-any-repo to be excluded to improve run time. This process makes heavy use of the search API, which has a very low quota that resets every minute. Expect to see many "waiting for quoto" interim messages.

Other work:
* Fix pre-commit python version for nb-ensure-clean. nb-ensure-clean requires python 3.6. A method in base64 is no longer supported in python 3.9