Fix "invalid token for org" bustage
It's now possible that a login can have org-owner access, but the
current PAT is not valid for that org.

We now use a javascript alert to notify the operator of each skipped
org.

Also added a framework to make it easier to notify operator of
initialization cell errors, even though those cells are
hidden-by-default. There are more places that framework could be added.