Fixed potential infinate loop
When an iterator makes a GitHub API call, that call could hit an API
limit. Therefore, you must process each returned item separately (so
that when you restart, you're in the same place), AND you must handle
the StopIteration exception yourself.

After that, forward progress is made on the iterator!

Prior code tried to slurp all the items into a list -- so the restart
was from the beginning of the query. Which almost guaranteed you'd hit
the limit again.